added CpuUsage and NetworkTransm Co-authored-by: jordanking7 <jordanking7@gmail.com>
